[Dev] Enable ThinLTO for CLI builds on Linux or OSX
LTO Makefile variable has been introduced earlier, should it be considered
for CLI environments?

On my non-scientific benchmarks ThinLTO is around 5% faster AND 15-20% smaller
on binary size. Only real downside is increase in linking time, and that's not
really tested with DuckDB (altough it should be mature and it's been used for a decade
by other complex projects, like browsers or DBMS).