It is currently possibly to alter the SQL semantics by creating a MACRO with the same name as an existing scalar (or aggregation) function.
MACROs get registered in the specified database/catalog and schema (or in the
default one). Putting them in the system catalog is prohibited already.

While stopping short of prohibiting defining MACROs with standard names, this PR
makes sure that the standard function is usedm whenever in SQL the function is not
explicitly qualified (by database and/or schema name); by prioritizing the default
system catalog for lookups of scalar functions.

A side-effect of this measure is that (extension) catalogs will get fewer lookups,
since every scalar/aggregation in a SQL query function prior to this PR was looked up
first in all catalogs (and default system catalog, where they reside, typically last).

(in MotherDuck, this lookup behavior then could lead to opening transactions that
 require RPCs even if the query itself were purely local)