GetWALIfExists: Avoid eagerly creating WAL file
This skips side effect of creating a WAL file on queries to GetWALSize.
This should have negligible effect, given in most cases this is delayed.

Workflow where this does makes a differences (altough negligible), is
for example performing CHECKPOINT after opening a database in read-write mode.

This codepath checks `GetWALLength`, but do not actually need to create a wal file.