Enable extensions test run with VS2019 C++ stdlib
This change enbales additional run of the test suite for all extensions
using C++ stdlib DLL from Visual Studion 2019. This is done to check
the compatibility with older versions of C++ stdlib introduced in #17991

New workflow step is only run for `x64-windows-static-md` build arch.

Testing: it was checked in manual CI runs, that added unit test run
actually causes a crash when `_DISABLE_CONSTEXPR_MUTEX_CONSTRUCTOR` is
not used.